Adding option to generate ArrayVector directly from a box mesh so that they are compatible.  Note this still needs testing

+++ src/utils/threadpool/AtomicList.hpp
@@ -28,7 +28,7 @@ AtomicList<TYPE, COMPARE>::AtomicList( size_t capacity,
       d_N_remove( 0 )
 {
     d_next[0] = -1;
-    for ( size_t i = 0; i < d_capacity; i++ ) {
+    for ( int i = 0; i < (int) d_capacity; i++ ) {
         d_next[i + 1] = -5 - i;
         d_objects[i]  = d_default;
     }
